Split SvgTextRef
