+++ Source/Text/SvgTextRef.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 
 namespace Svg
 {
@@ -13,27 +11,6 @@ namespace Svg
             get { return GetAttribute<Uri>("href", false); }
             set { Attributes["href"] = value; }
         }
-
-#if !NO_SDC
-        internal override IEnumerable<ISvgNode> GetContentNodes()
-        {
-            var refText = this.OwnerDocument.IdManager.GetElementById(this.ReferencedElement) as SvgTextBase;
-            IEnumerable<ISvgNode> contentNodes = null;
-
-            if (refText == null)
-            {
-                contentNodes = base.GetContentNodes();
-            }
-            else
-            {
-                contentNodes = refText.GetContentNodes();
-            }
-
-            contentNodes = contentNodes.Where(o => !(o is ISvgDescriptiveElement));
-
-            return contentNodes;
-        }
-#endif
 
         public override SvgElement DeepCopy()
         {
